// web/lib/api/generated/model/filter.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { FilterKindEnum } from './filterKindEnum';

export interface Filter {
  key: string;
  label: string;
  kind: FilterKindEnum;
  required: boolean;
  choices: string[];
  /** @nullable */
  default: string | null;
  entity: string;
}